Make checkpoint truncating happen during report creation
