+++ src/main/java/nl/nn/testtool/TestTool.java
@@ -169,11 +169,7 @@ public class TestTool {
 					message = report.checkpoint(threadId, sourceClassName, name, message, checkpointType, levelChangeNextCheckpoint);
 					reportsInProgressEstimatedMemoryUsage = reportsInProgressEstimatedMemoryUsage + report.getEstimatedMemoryUsage();
 					if (report.finished()) {
-						report.setEndTime(System.currentTimeMillis());
-						if(maxMessageLength > 0) {
-							truncateCheckpointMessages(report);
-						}
-						
+						report.setEndTime(System.currentTimeMillis());						
 						log.debug("Report is finished for '" + correlationId + "'");
 						reportsInProgress.remove(report);
 						reportsInProgressByCorrelationId.remove(correlationId);
@@ -190,29 +186,6 @@ public class TestTool {
 			}
 		}
 		return message;
-	}
-
-	private void truncateCheckpointMessages(Report report) {
-		// For a message that is referenced by multiple checkpoints, have one truncated message that is
-		// referenced by those checkpoints, to prevent creating multiple String objects representing the
-		// same string and occupying unnecessary memory.
-		Map<String, String> truncatedMessages = new RefCompareMap<String, String>();
-		
-		for(Checkpoint cp : report.getCheckpoints()) {
-			if(cp.getMessage() != null) {				
-				if(truncatedMessages.containsKey(cp.getMessage())) {
-					cp.setPreTruncatedMessageLength(cp.getMessage().length());
-					cp.setMessage(truncatedMessages.get(cp.getMessage()));
-				} else if(cp.getMessage().length() > maxMessageLength) {
-					String truncatedMessage = cp.getMessage().substring(0, maxMessageLength)
-						+ "... ("+(cp.getMessage().length() - maxMessageLength)+" more characters)";
-					
-					truncatedMessages.put(cp.getMessage(), truncatedMessage);
-					cp.setPreTruncatedMessageLength(cp.getMessage().length());
-					cp.setMessage(truncatedMessage);
-				}
-			}
-		}
 	}
 	
 	public Object startpoint(String correlationId, String sourceClassName, String name, Object message) {
@@ -419,92 +392,5 @@ public class TestTool {
 
 	public static String getImplementationVersion() {
 		return Package.getPackage("nl.nn.testtool").getImplementationVersion();
-	}
-}
-
-/**
- * A custom implementation of the map interface that compares keys based on their object reference, i.e. comparing with 'o1 == o2' rather than 'o1.equals(o2)'.
- * This greatly enhances the performance of maps with large objects as keys.
- * <p>
- * Note: Since this implementation was written with a specific use case in mind, most methods are not implemented and will throw an exception when called. 
- */
-class RefCompareMap<K, V> implements Map<K, V> {
-	
-	private List<K> keys = new ArrayList<K>();
-	private List<V> values = new ArrayList<V>();
-	
-	@Override
-	public V get(Object key) {
-		int i = 0;
-		for(Object o : keys) {
-			if(o == key) {
-				return values.get(i);
-			}
-			i++;
-		}
-		return null;
-	}
-	
-	@Override
-	public V put(K key, V value) {
-		keys.add(key);
-		values.add(value);
-		
-		return value;
-	}
-	
-	@Override
-	public boolean containsKey(Object key) {
-		for(Object o : keys) {
-			if(o == key) {
-				return true;
-			}
-		}
-		return false;
-	}
-
-	@Override
-	public void clear() {
-		throw new NotImplementedException();
-	}
-
-	@Override
-	public boolean containsValue(Object value) {
-		throw new NotImplementedException();
-	}
-
-	@Override
-	public Set entrySet() {
-		throw new NotImplementedException();
-	}
-
-	@Override
-	public boolean isEmpty() {
-		throw new NotImplementedException();
-	}
-
-	@Override
-	public Set keySet() {
-		throw new NotImplementedException();
-	}
-
-	@Override
-	public void putAll(Map m) {
-		throw new NotImplementedException();
-	}
-
-	@Override
-	public V remove(Object key) {
-		throw new NotImplementedException();
-	}
-
-	@Override
-	public int size() {
-		throw new NotImplementedException();
-	}
-
-	@Override
-	public Collection values() {
-		throw new NotImplementedException();
 	}
 }
